+++ src/Router/MimeType.ts
@@ -1,0 +1,14 @@
+enum MimeType {
+  IMAGE_JPEG = "image/jpeg",
+  IMAGE_GIF = "image/gif",
+  IMAGE_PNG = "image/png",
+  IMAGE_SVG = "image/svg+xml",
+  IMAGE_WEBP = "image/webp",
+  PLAIN_TEXT = "text/plain",
+  JSON = "application/json",
+  HTML = "text/html",
+  JAVASCRIPT = "text/javascript",
+  FORM_DATA = "multipart/form-data",
+}
+
+export default MimeType;